requirements.txt
aiohttp==3.7.4.post0
aiohttp-jinja2==1.5
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
filelock==3.0.12
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
Pillow==8.3.2
pipenv==2020.8.13
pycosat==0.6.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
yarl==1.6.3
